feat: Add Session 5 Admit and Titanic modules
This commit introduces two new modules, `Admit` and `Titanic`, to
Session 5. These modules implement multi-layer perceptrons for
university admission prediction and Titanic survival prediction,
respectively.

The changes include:
- Adding new Haskell source files for `Admit` and `Titanic` models.
- Implementing data extraction and preprocessing logic specific to each
  dataset.
- Configuring MLP hyperparameters, loss functions, and training
  parameters.
- Integrating with existing evaluation and charting utilities.
- Updating `.gitignore` to include `Session5/data/`.
- Adding a comprehensive `README.md` for Session 5 detailing
  classification metrics and analysis of the implemented models.
- Adding new Haskell source files for `Evaluation` and `LoadData`
  utilities.
- Modifying `internship.cabal` and `package.yaml` to include the new
  source directories and modules.

+++ Session4/MlpXor.hs
@@ -12,7 +12,7 @@ numIters = 500
 learningRate :: Tensor
 learningRate = 1e-1
 
-trainingData :: [([Float], Float)]
+trainingData :: [([Float], Int)]
 trainingData = [ ([0, 0], 0)
                , ([0, 1], 1)
                , ([1, 0], 1)
